Guarantee access is denied if not authorized
Note on how this impacts the gc (guiding cancer) app:

Something in RestDefaultsController causes a redirect to the login page
even if the definition of role_authorization_failure_response in
AuthorizesWithRoles does not render or redirect.

But if you include AuthorizesWithRoles in a controller, you don't include
RestDefaultsController, and role_authorization_failure_response does not
render or redirect, then the filter chain will not be halted and the
request will be served as if the user is authorized. That's one case that
makes this commit worth doing.